Skeleton of resolving and applying labels copy-pasted from github-bot
With a few tweaks related to:

- Logging: using `@actions/core` methods like `core.debug()` instead of
		   `bunyan` like the github-bot has been using
- GitHub Client: using `@actions/github` which is more or less exactly
			     what the github-bot also used, namely the
			     `oktokit/rest.js` package, it has just been pieced
			     together to be more convenient for GitHub Actions code

Other than that, it's copy-pasting of what was found in the github-bot
project:

- `./lib/node-repo.js`
- `./lib/node-labels.js` (renamed to `resolve-labels.js`)

..removing code that was not related to resolving & applying labels, along
with the related tests.

Currently this is *not* to be considered a GitHub Action, it's only the
said files from github-bot, adjusted slightly to the `@actions/*`
packages as a precursor to adding GitHub Action artifacts like
`action.yml` and so on.

Copied from this specific point in time: [nodejs/github-bot 737094](https://github.com/nodejs/github-bot/tree/737094cf07224be73eafa279356fd44a1cc55c87)