Extract filepaths -> labels rules from code into external .yml config
First successful step in moving the rules between filepaths and label
names from the .js source code, into an external .yml configuration
file.

At the moment, the .yml location is hard coded to just make the test
suite pass as before.

That won't be of much help in the long run, as GitHub repos will want to
be configure the .yml location themselves and ofcourse for them to have
that file in their own repository.

Hence the next step will be to make the path configurable and download
the configuration file contents from the target repo, where this
GitHub Action is destined to run.