Remove waiting for 2 seconds before resolving labels
In the github-bot we have found the need to wait a couple of seconds
before resolving labels on newly opened PRs.

That's because we stumbled upon what felt like consistency issues on
GitHub's side, where the github-bot got a webhook event about an opened
PR, its API would in some scenarios return 404 when fetching existing
labels on that PR.

Waiting a couple of seconds did the trick.

When the same code runs in a GitHub Action, we'd that same effect of not
asking too early out of the box, because it takes a while for Actions
to boot up and do their thing.

Refs https://github.com/nodejs/github-bot/issues/79