Provide external configuration file contents when resolving labels
These changes makes our resolve labels logic unaware of *where* the
configuration / label rules comes from. As long as the .yml contents
is provided to `.resolveLabels()` as a `string`, it's all fine and
dandy.

This is exactly what we'll need around the corner when truly making
this a GitHub Action, where it's planned to fetch this configuration
from the target repo somewhere..